Add setuptools to github action Build Wheel

+++ scripts/python/build_exe.py
@@ -18,8 +18,23 @@ import os
 import pathlib
 import shutil
 import zipfile
+import platform
+import subprocess
 
 from cx_Freeze import Freezer, Executable
+
+
+def get_mac_architecture():
+  try:
+    # Get the hardware architecture using sysctl
+    arch = subprocess.check_output(
+      ['sysctl', '-n', 'hw.machine'],
+      stderr=subprocess.DEVNULL
+    ).decode().strip()
+    return arch
+  except subprocess.CalledProcessError:
+    # Fallback to platform.machine() if sysctl isn't available (unlikely on macOS)
+    return platform.machine()
 
 FILE_ROOT_PATH = pathlib.Path(__file__).parent
 
@@ -59,29 +74,33 @@ def remove_dist_info_folders(directory: pathlib.Path):
 
 
 if __name__ == '__main__':
+  tmp_python_version = f"{sys.version_info.major}.{sys.version_info.minor}"
+  tmp_shared_suffix = f".cpython-{tmp_python_version.replace('.', '')}-darwin.so"
+  tmp_arch = get_mac_architecture()
+  tmp_build_dir_name = f"exe.macosx-{tmp_arch}-{tmp_python_version}"
   freezer.freeze()
-  with zipfile.ZipFile(pathlib.Path(f"{FILE_ROOT_PATH}/build/exe.linux-x86_64-3.11/lib/library.zip"), 'r') as zip_ref:
-    zip_ref.extractall(pathlib.Path(f"{FILE_ROOT_PATH}/build/exe.linux-x86_64-3.11/lib"))
-  _CMD_FROM_BUILD_DIR = pathlib.Path(FILE_ROOT_PATH.parent / "cmake-build-release" / "_cmd.cpython-311-x86_64-linux-gnu.so")
-  _CMD_FROM_PRE_BUILT_DIR = pathlib.Path(FILE_ROOT_PATH.parent / "pre-built" / "_cmd.cpython-311-x86_64-linux-gnu.so")
-  if _CMD_FROM_BUILD_DIR.exists():
-    shutil.copy(
-      _CMD_FROM_BUILD_DIR,
-      pathlib.Path(FILE_ROOT_PATH / "build/exe.linux-x86_64-3.11/lib/pymol" / "_cmd.cpython-311-x86_64-linux-gnu.so")
-    )
+  with zipfile.ZipFile(pathlib.Path(f"{FILE_ROOT_PATH}/build/{tmp_build_dir_name}/lib/library.zip"), 'r') as zip_ref:
+    zip_ref.extractall(pathlib.Path(f"{FILE_ROOT_PATH}/build/{tmp_build_dir_name}/lib"))
+  if not pathlib.Path(const.PROJECT_ROOT_DIR / "src/python/pymol").exists():
+    utils.copy_pymol_sources()
+    _CMD_FROM_BUILD_DIR = pathlib.Path(const.PROJECT_ROOT_DIR / "cmake-build-release" / f"_cmd{tmp_shared_suffix}")
+    if _CMD_FROM_BUILD_DIR.exists():
+      shutil.copy(
+        _CMD_FROM_BUILD_DIR,
+        pathlib.Path(const.PROJECT_ROOT_DIR / "src/python/pymol" / f"_cmd{tmp_shared_suffix}")
+      )
+    else:
+      print(f"Could not find _cmd{tmp_shared_suffix} for building the EXE file.")
   else:
-    shutil.copy(
-      _CMD_FROM_PRE_BUILT_DIR,
-      pathlib.Path(FILE_ROOT_PATH / "build/exe.linux-x86_64-3.11/lib/pymol" / "_cmd.cpython-311-x86_64-linux-gnu.so")
-    )
-  remove_dist_info_folders(pathlib.Path(FILE_ROOT_PATH / "build/exe.linux-x86_64-3.11/lib"))
+    print("The src/python/pymol directory already exists, that might mean a self compiled _cmd module was built.")
+  remove_dist_info_folders(pathlib.Path(FILE_ROOT_PATH / f"build/{tmp_build_dir_name}/lib"))
   shutil.copytree(
     str(pathlib.Path(FILE_ROOT_PATH / "pymol/wizard")),
-    str(pathlib.Path(FILE_ROOT_PATH / "build/exe.linux-x86_64-3.11/lib/pymol/wizard")),
+    str(pathlib.Path(FILE_ROOT_PATH / f"build/{tmp_build_dir_name}/lib/pymol/wizard")),
     dirs_exist_ok=True
   )
   shutil.copytree(
     str(pathlib.Path(FILE_ROOT_PATH / "pymol/data/startup")),
-    str(pathlib.Path(FILE_ROOT_PATH / "build/exe.linux-x86_64-3.11/lib/pymol/data/startup")),
+    str(pathlib.Path(FILE_ROOT_PATH / f"build/{tmp_build_dir_name}/lib/pymol/data/startup")),
     dirs_exist_ok=True
   )
